01-01-2016 17:18
- wyswietlanie obserwowanych
- dodawanie do obserwowanych tylko dla zalogowanych uzytkownikow
- zablokowana możliwosć usuwania nieswoich ofert
- usuniecie z listy obserwowanych po usunieciu oferty

+++ Bootstrap/src/AppBundle/Controller/UsunOferteController.php
@@ -8,6 +8,8 @@ use AppBundle\Entity\User;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
+use FOS\UserBundle\Model\UserInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class UsunOferteController extends Controller
 {
@@ -15,11 +17,18 @@ class UsunOferteController extends Controller
      * @Route("/oferta/{idOferty}/delete", name="App_UsunOferte")
      */
     public function deletebyAction($idOferty){
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
 
         $em = $this->getDoctrine()->getEntityManager();
         $oferta = $em->getRepository('AppBundle:Oferty')->findOneBy(
             array('idOferty' => $idOferty)
         );
+        if($oferta->getUserId()!=$user){
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
 
         $em->remove($oferta);
         $em->flush();
